+++ incl/Favorites.hpp
@@ -3,13 +3,17 @@
 using namespace geode::prelude;
 
 namespace favorites {
-    extern Mod* m_thisMod; // Favorite Mods
+    extern Mod* m_thisMod = Mod::get(); // Favorite Mods
 
-    // Check to see if a mod is included in the user's favorites
-    bool isFavorite(
-        std::string modId // Your mod's ID
-    );
+    /**
+     * Check to see if a mod is included in the user's favorites
+     *
+     * @param modID Your mod's ID
+     */
+    bool isFavorite(std::string modID);
 
-    // Get the amount of mods the player has in their favorites
+    /**
+     * Get the amount of mods the player has in their favorites
+     */
     int favoritesCount();
 };