+++ incl/src/Favorites.cpp
@@ -2,15 +2,11 @@
 
 #include <Geode/Geode.hpp>
 
-#include <Geode/utils/terminate.hpp>
-
 using namespace geode::prelude;
 
 namespace favorites {
-    Mod* m_thisMod = Mod::get();
-
-    bool isFavorite(std::string modId) {
-        return m_thisMod->getSavedValue<bool>(modId);
+    bool isFavorite(std::string modID) {
+        return m_thisMod->getSavedValue<bool>(modID);
     };
 
     int favoritesCount() {